archive/v1/scripts/requirements.txt
# Core scraping and browser automation
selenium==4.8.0
beautifulsoup4==4.12.2
requests==2.31.0

# Data processing and handling
pandas==1.5.3
numpy==1.24.2

# Database interaction
sqlalchemy==2.0.5
psycopg2-binary==2.9.5  # For PostgreSQL (adjust for other DBs)

# Additional utilities
lxml==4.9.3
python-dotenv==1.0.0  # For environment variable handling

# Testing
pytest==7.4.0